requirements.txt
asgiref==3.8.1
Django==5.1.6
django-appconf==1.1.0
django-cf==0.0.17
django-compressor==4.5.1
django-tinymce==4.1.0
pillow==11.1.0
python-dotenv==1.0.1
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
websocket-client==1.8.0
whitenoise==6.9.0
